getTempHum ohne WiringPi.h, main, getPres geändert
getPres(): Berechnung von Luftdruck auf Meeresniveau geändert. Trotzdem
sind die Messwerte zu klein!

main(): fprintf anstatt printf verwenden.  Rückgabewert von getTempHum
wird geprüft.

common.h: doSleep() Funktion eingefügt, weil ich diese auch in
getTempHum verwenden will

einige #define angelegt, damit der Code cooler ausschaut!
spezielle Datentypen eingeführt: uint8_t uint16_t etc. damit der Code
mehr nach Embedded System ausschaut (=noch cooler)

+++ writeLED.c
@@ -1,7 +1,7 @@
 #include "common.h"
 #include <stdio.h>
 
-extern void writeLED(int mode)
+extern void writeLED(uint8_t mode)
 {
 	mode=mode;
 	fprintf(stdout, "writeLED not yet implemented!\n");
